+++ src/T3MP/BenchmarkMode.cs
@@ -1,7 +1,0 @@
-namespace T3MP;
-
-internal enum BenchmarkMode
-{
-    Vanilla = 0,
-    Optimized = 1
-}